requirements.txt
sphinx_rtd_theme
gnutools-python
